requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.4.0
requests>=2.32.0
torch>=2.3.0
sentence-transformers>=3.0.0
transformers>=4.44.0
easyocr>=1.7.0
faiss-cpu>=1.8.0
Pillow>=10.0.0

google-generativeai>=0.7.0

